+++ src/xSTIR/cSTIR/tests/getenv.h
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace sirf {
-	std::string getenv(const char* name, bool throws=false);
-}