+++ src/xSTIR/pSTIR/STIR.py
@@ -59,6 +59,33 @@ MAX_ACQ_DIMS = 10
 MAX_IMG_DIMS = 10
 
 
+def get_STIR_version_string():
+    """Returns STIR engine version as Python str."""
+    handle = pystir.cSTIR_STIR_version_string()
+    check_status(handle)
+    version = pyiutil.charDataFromHandle(handle)
+    pyiutil.deleteDataHandle(handle)
+    return version
+
+
+def get_STIR_doc_dir():
+    """Returns STIR engine version as Python str."""
+    handle = pystir.cSTIR_get_STIR_doc_dir()
+    check_status(handle)
+    path = pyiutil.charDataFromHandle(handle)
+    pyiutil.deleteDataHandle(handle)
+    return path
+
+
+def get_STIR_examples_dir():
+    """Returns STIR engine version as Python str."""
+    handle = pystir.cSTIR_get_STIR_examples_dir()
+    check_status(handle)
+    path = pyiutil.charDataFromHandle(handle)
+    pyiutil.deleteDataHandle(handle)
+    return path
+
+
 def set_verbosity(verbosity):
     """Set the verbosity of all STIR output."""
     try_calling(pystir.cSTIR_setVerbosity(verbosity))